:bug:: fix: userRepository id type Integer to Long

+++ src/main/java/com/grimeet/grimeet/domain/user/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.grimeet.grimeet.domain.user.repository;
 
-import com.grimeet.grimeet.domain.user.dto.UserResponseDto;
 import com.grimeet.grimeet.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     /**
      * 이메일로 유저 정보를 조회하는 메서드
      * @param email
